+++ PI3FernandezTrigoAlejandro/ejercicios/ejercicio4recursivaconmemoria.h
@@ -17,7 +17,7 @@
 #include "types/hash_table.h"
 
 // Cabeceras de las funciones:
-long ejercicio4RecursivoConMemoria (int numeroA, int numeroB);
-long ejercicio4RecursivoConMemoriaInterno (int numeroA, int numeroB, hash_table memoria);
+int ejercicio4RecursivoConMemoria (int numeroA, int numeroB);
+int ejercicio4RecursivoConMemoriaInterno (int numeroA, int numeroB, hash_table memoria);
 
 #endif /* EJERCICIOS_EJERCICIO4RECURSIVACONMEMORIA_H_ */
